modules/locations-api/requirements.txt
alembic==1.7.5
certifi==2021.5.30
click==8.0.2
Flask==2.0.2
GeoAlchemy2==0.10.0
greenlet==1.1.2
grpcio==1.40.0
grpcio-tools==1.40.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
kafka-python==2.0.2
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
packaging==21.3
protobuf==3.18.0
psycopg2-binary==2.9.3
pyparsing==3.0.6
Shapely==1.8.0
six==1.16.0
sqla-wrapper==5.2
SQLAlchemy==1.4.29
Werkzeug==2.0.2
zipp==3.7.0